Extract out functionality and move into its own type.
